Fix dependency conflict when apps use support-v4
When an app wants to use the latest support-v4 library and the libnetcipher library, they have a problem when using gradle. When trying to build, gradle runs into a conflict where the libnetcipher library is supplying its own support-v4 jar. This problem could be overridden using the "exclude module" gradle command (see https://stackoverflow.com/questions/20989317/multiple-dex-files-define-landroid-support-v4-accessibilityservice-accessibility) but it currently does not work since libnetcipher is explicitly including a jar and that cannot be overridden