purge unused code to eliminate dependency on android-support-v4.jar
The only code that still used android-support-v4.jar was marked by Eclipse
as unused.  Also, it was about building a Notification, which should be
handled by the app anyway, not a library.  If a library needs to manage a
Notification, it should be created by the app and passed into the library.

closes https://github.com/guardianproject/NetCipher/pull/20