modernize download Intent and choose between FDroid, Play, and f-droid.org
This fixes up the Intent used to take the user to install Orbot.  First,
it looks for Google Play or FDroid, and if neither are installed, then it
takes the user to the f-droid.org page for Orbot.  Google Play and F-Droid
are our only supported download methods for Orbot.  Taking users to other
markets, like MiMarket, Baidu, Amazon, etc. will just give them a confusing
blank result.

closes #408 https://dev.guardianproject.info/issues/408