convert all methods without instance state to static utility methods
Since the only state that is stored in a OrbotHelper instance is the
Context that was passed in, for most of the methods, there is no need to
instantiate OrbotHelper to use them.  All of the methods that have a
Context or Activity definitely do not need the stored Context. Static
methods that are passed a Context are a more common pattern these days.

This change should be fully backwards compatible.  For the methods that
changed to static, the app will just have warnings saying that the static
call is preferred over a call on an instance.