move gpg-signing out of make-release-build to speed up process
Using faketime makes the build process quite slow, since some parts rely on
clock to tick in order to progress.  This also moves the GPG signing
to ./make-release-build since it does not need faketime, and it is one of
the things that requires the clock to tick before moving on.