getHttpURLConnection: if useTor() is enabled, disable setting global proxy
Traffic must be directed to Tor using the proxy settings, and Orbot has its
own proxy settings for connections that need proxies to work.  So if "use
Tor" is enabled, as tested by looking for the static instance of Proxy,
then no other proxy settings are allowed to override the current Tor proxy.