travis-ci: also run tests in android-10 emulator
Since a big part of netcipher is improving HTTPS on old devices, it should
be tested on android-10 to make sure it is actually doing that. android-16
is when they made some big improvements to the HTTPS stuff, so this should
test against something older than that.  android-10 is a common min SDK
since it is the oldest version of Android that is still included on new
devices.

This also tests the Orbot installed check by running the first emulator
without Orbot and the second emulator with Orbot installed.