URLConnection test: remove goo.gl, add web.wechat.com & f-droid.org
The tests were failing to complete a connection against goo.gl, while all
the rest were working fine.  I replaced goo.gl with f-droid.org since it is
a host that is used with NetCipher and is well maintained with a high
uptime.  web.wechat.com is a big Chinese service that uses HTTPS