move sub-lib dirs to match the lib names
gradle will automatically use the dirname as the project name, and then
also name the release JARs after that project name.  So rename the dirs of
the libraries that have their release builds built by gradle.  The release
builds of the core 'netcipher' library are still built with ant, so that
dir is still called 'libnetcipher'.