HttpURLConnection: disable TLS Session Tickets
TLS Session Tickets are used to resume TLS sessions.  They are sent in
plain text before TLS is started up.  They are in effect a cookie that can
be viewed from anyone who can see the network traffic.  Even worse, the
server can easily set specific cookies in order to track users.  They are
used to speed up renegotiation of TLS on reconnect, so disabling this will
cause increased latency and bandwidth usage.

Thanks to @dkg for alerting me to this issue.

https://timtaubert.de/blog/2014/11/the-sad-state-of-server-side-tls-session-resumption-implementations/