temporary disable broken unit tests
Test are broken because the they depend on a file hosted on the web
which is not longer available.  This commit disables all affected tests
and puts the missing file on gitlab.com.  It also changes all test urls
to the upcomming gitlab test url.  In a followup commit they can be
reenable and will start working as expected again.