bump to gradle 4.4.1 and Gradle Android Plugin 3.1.4
Since some deps of the webkit stuff require Java8, we have to do this. The
gradle v4.4.1 was chosen because it is built reproducibly in Debian/buster
and Gradle Android Plugin 3.1.4 is the newest plugin that works with that
version of gradle.

gradle checksum from https://gradle.org/release-checksums/

This also bumped the minSdkVersion to 9 because of:
    Execution failed for task ':libnetcipher:processDebugAndroidTestManifest'.
    > Manifest merger failed : uses-sdk:minSdkVersion 8 cannot be smaller than version 9 declared in library [com.android.support.test:rules:1.0.2] /builds/eighthave/NetCipher/.gradle/caches/transforms-1/files-1.1/rules-1.0.2.aar/5550aaa96d5a502611cb40b5b1e82ab5/AndroidManifest.xml as the library might be using APIs not available in 8
      	Suggestion: use a compatible library with a minSdk of at most 8,
      		or increase this project's minSdk version to at least 9,
      		or use tools:overrideLibrary="android.support.test" to force usage (may lead to runtime failures)