useGlobalProxyWithDNSLeaksOnAndroid7x for global proxying despite DNS leak
The global proxying seems to work fine on 7.x, just the DNS skips the proxy.
So this provides a method to enable global proxying on 7.x for people who
need just access, not privacy.  It is also useful as a failsafe for manual
proxy setups, since it applies globally to the app.  Leaking only DNS is
better than leaking all traffic.